+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Allow `MarkdownField` to emit value change event (`onValueChange` component's attribute).
 - `FileField` shows the files given as value. The `value` attribute was ignored before.
 - `FileField` is correctly translated. Resolve a bug that only displays translation key.
+- Error messages from UiKit are now correctly translated.
 
 ## [v0.4.0] - 2020-09-10
 ### Added