+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html) since [v0.1.0].
 
 ## [Unreleased]
+
+## [v0.4.1] - 2020-09-14
 ### Fixed
 - Allow `MarkdownField` to emit value change event (`onValueChange` component's attribute).
 - `FileField` shows the files given as value. The `value` attribute was ignored before.
@@ -342,6 +344,7 @@ be rendered as an `a`, a `button` or a `Link` from `react-router-dom`
 - Setup storybook as documentation
 
 [Unreleased]: https://github.com/ferpection/uikit/tree/master
+[v0.4.1]: https://github.com/ferpection/uikit/tree/v0.4.1
 [v0.4.0]: https://github.com/ferpection/uikit/tree/v0.4.0
 [v0.3.2]: https://github.com/ferpection/uikit/tree/v0.3.2
 [v0.3.1]: https://github.com/ferpection/uikit/tree/v0.3.1